Moved comms code into separate file
